requirements.txt
google-generativeai==0.8.6
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.104.0
uvicorn==0.24.0
pytest==7.4.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
pillow==10.1.0
python-multipart==0.0.6